refactor(datenraum): split components
